Move post component to post module
