+++ client/modules/post/components/info-display/details-info.tsx
